requirements.txt
ultralytics
fastapi
uvicorn
pyyaml
